Modify the glob transformation logic to also strip trailing slashes
