+++ packages/steiger/src/models/config/transform-globs.ts
@@ -1,9 +1,10 @@
 import { posix, sep } from 'node:path'
+import { pipe } from 'ramda'
 import { Config } from '@steiger/types'
 
 import { isConfigObject, isConfiguration } from './raw-config'
 
-function convertGlobs(rootPath: string, globs: Array<string>) {
+function convertRelativeGlobsToAbsolute(rootPath: string, globs: Array<string>) {
   function composeAbsolutePath(root: string, glob: string) {
     // Remove '/'. The root has platform-specific separators
     const segmentsOfRoot = root.slice(1).split(sep)
@@ -14,10 +15,18 @@ function convertGlobs(rootPath: string, globs: Array<string>) {
   return globs.map((glob) => (glob.startsWith('.') ? composeAbsolutePath(rootPath, glob) : glob))
 }
 
-export function convertRelativeGlobsToAbsolute(config: Config, configLocationPath: string | null) {
+function stripTrailingSlashes(globs: Array<string>) {
+  return globs.map((ignore) => ignore.replace(/\/$/, ''))
+}
+
+export function transformGlobs(config: Config, configLocationPath: string | null) {
   if (!configLocationPath) {
     return config
   }
+
+  const globsTransformationPipeline = pipe(stripTrailingSlashes, (globs) =>
+    convertRelativeGlobsToAbsolute(configLocationPath, globs),
+  )
 
   return config.map((item) => {
     if (!isConfiguration(item)) {
@@ -29,11 +38,11 @@ export function convertRelativeGlobsToAbsolute(config: Config, configLocationPat
     }
 
     if (newItem.ignores) {
-      newItem.ignores = convertGlobs(configLocationPath, newItem.ignores)
+      newItem.ignores = globsTransformationPipeline(newItem.ignores)
     }
 
     if (isConfigObject(newItem) && newItem.files) {
-      newItem.files = convertGlobs(configLocationPath, newItem.files)
+      newItem.files = globsTransformationPipeline(newItem.files)
     }
 
     return newItem
